requirements.txt
fastapi
sqlmodel
uvicorn
pyodbc
pandas
openpyxl
python-multipart
pyjwt
passlib[bcrypt]
python-dotenv
pydantic[email]